+++ KmpIO/src/commonMain/kotlin/com/oldguy/common/io/Directory.kt
@@ -66,6 +66,8 @@ class Directory(dirPath: String)
                     throw IllegalStateException("Attempt to delete ${this[i].fullPath} failed")
             }
         }
+        if (!directory.delete())
+            throw IllegalStateException("Attempt to delete ${directory.fullPath} failed")
     }
 
     companion object {